Ultima version
ESTA VEZ SI QUE LO TOQUE POR TODOOOOOOS LADOOOS!. Hay que revisar las
funcionalidades para que siga andando todo OK. Ya son casi las 2...
Cualquier cosa avisame.

+++ VisualStudio/VS/Entidad/Venta.cs
@@ -14,7 +14,7 @@ namespace VisualStudio.Entidad
         public string EmailComprador { set; get; }
         public DateTime FechaTransaction { set; get; }
         public int Cantidad { set; get; }
-        public int PrecioUnitario { set; get; }
+        public float PrecioUnitario { set; get; }
         public string Estado { set; get; }
     }
 }
